API Refresh

+++ addons/godotsteam_csharpbindings/Steam.HTTP.Signals.cs
@@ -5,9 +5,9 @@ namespace GodotSteam;
 
 public static partial class Steam
 {
-    public delegate void HttpRequestCompletedEventHandler(long cookieHandle, long contextValue, bool requestSuccess, long statusCode, long bodySize);
+    public delegate void HttpRequestCompletedEventHandler(uint cookieHandle, ulong contextValue, bool requestSuccess, int statusCode, uint bodySize);
     private static event HttpRequestCompletedEventHandler HttpRequestCompletedEvent;
-    static Action<long, long, bool, long, long> _httpRequestCompletedAction = (cookieHandle, contextValue, requestSuccess, statusCode, bodySize) =>
+    static Action<uint, ulong, bool, int, uint> _httpRequestCompletedAction = (cookieHandle, contextValue, requestSuccess, statusCode, bodySize) =>
     {
         HttpRequestCompletedEvent?.Invoke(cookieHandle, contextValue, requestSuccess, statusCode, bodySize);
     };
@@ -31,9 +31,9 @@ public static partial class Steam
         }
     }
     
-    public delegate void HttpRequestDataReceivedEventHandler(long cookieHandle, long contextValue, long offset, long bytesReceived);
+    public delegate void HttpRequestDataReceivedEventHandler(uint cookieHandle, ulong contextValue, uint offset, uint bytesReceived);
     private static event HttpRequestDataReceivedEventHandler HttpRequestDataReceivedEvent;
-    static Action<long, long, long, long> _httpRequestDataReceivedAction = (cookieHandle, contextValue, offset, bytesReceived) =>
+    static Action<uint, ulong, uint, uint> _httpRequestDataReceivedAction = (cookieHandle, contextValue, offset, bytesReceived) =>
     {
         HttpRequestDataReceivedEvent?.Invoke(cookieHandle, contextValue, offset, bytesReceived);
     };
@@ -57,9 +57,9 @@ public static partial class Steam
         }
     }
     
-    public delegate void HttpRequestHeadersReceivedEventHandler(long cookieHandle, long contextValue);
+    public delegate void HttpRequestHeadersReceivedEventHandler(uint cookieHandle, ulong contextValue);
     private static event HttpRequestHeadersReceivedEventHandler HttpRequestHeadersReceivedEvent;
-    static Action<long, long> _httpRequestHeadersReceivedAction = (cookieHandle, contextValue) =>
+    static Action<uint, ulong> _httpRequestHeadersReceivedAction = (cookieHandle, contextValue) =>
     {
         HttpRequestHeadersReceivedEvent?.Invoke(cookieHandle, contextValue);
     };
